updated RatingSelector
