+++ src/components/App/views/PageStaffSupport/PageStaffSupport.js
@@ -4,24 +4,11 @@ import NextButton from "../../../NextButton/NextButton";
 import RatingSelector from "../../../RatingSelector/RatingSelector";
 
 class PageStaffSupport extends Component {
-  handleSelection = (event) => {
-    this.props.dispatch({
-      type: "MODIFY_NEW_FEEDBACK",
-      payload: {
-        property: "support",
-        value: event.target.value,
-      },
-    });
-  };
-
   render() {
     return (
       <div>
         <h2>Staff Support</h2>
-        <RatingSelector
-          handleSelection={this.handleSelection}
-          value={this.props.store.newFeedbackReducer.support}
-        />
+        <RatingSelector property="support" />
         <NextButton nextPath="/comments" />
       </div>
     );
