+++ src/components/App/views/PageUnderstanding/PageUnderstanding.js
@@ -4,24 +4,11 @@ import NextButton from "../../../NextButton/NextButton";
 import RatingSelector from "../../../RatingSelector/RatingSelector";
 
 class PageUnderstanding extends Component {
-  handleSelection = (event) => {
-    this.props.dispatch({
-      type: "MODIFY_NEW_FEEDBACK",
-      payload: {
-        property: "understanding",
-        value: event.target.value,
-      },
-    });
-  };
-
   render() {
     return (
       <div>
         <h2>Understanding</h2>
-        <RatingSelector
-          handleSelection={this.handleSelection}
-          value={this.props.store.newFeedbackReducer.understanding}
-        />
+        <RatingSelector property="understanding" />
         <NextButton nextPath="/support" />
       </div>
     );
