+++ src/components/RatingSelector/RatingSelector.js
@@ -1,10 +1,21 @@
 //  selector 1-5
 import React, { Component } from "react";
+import { connect } from "react-redux";
 
 class RatingSelector extends Component {
+  handleSelection = (event) => {
+    this.props.dispatch({
+      type: "MODIFY_NEW_FEEDBACK",
+      payload: {
+        property: this.props.property,
+        value: event.target.value,
+      },
+    });
+  };
+
   render() {
     return (
-      <select onChange={this.props.handleSelection} value={this.props.value}>
+      <select onChange={this.handleSelection} value={this.props.value}>
         <option value="">Please select 1-5</option>
         <option value="1">1</option>
         <option value="2">2</option>
@@ -16,4 +27,8 @@ class RatingSelector extends Component {
   }
 }
 
-export default RatingSelector;
+const putStoreOnProps = (store) => ({
+  store,
+});
+
+export default connect(putStoreOnProps)(RatingSelector);
